requirements.txt
flask>=3.0.0
primer3-py>=2.0.0
requests>=2.31.0
